requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiomysql==0.2.0
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
APScheduler==3.10.4
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
blosc2==2.0.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.14
click==8.1.7
coloredlogs==15.0.1
cryptography==41.0.3
Cython==3.0.2
dataclasses-json==0.5.14
decorator==5.1.1
deprecation==2.1.0
dnspython==2.4.2
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.103.2
fastapi_amis_admin==0.6.7
filelock==3.12.4
financetoolkit==1.3.5
finnhub-python==2.4.19
flatbuffers==23.5.26
frozendict==2.3.8
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.35
greenlet==2.0.2
grpcio==1.59.0
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
langchain==0.0.286
langsmith==0.0.35
loguru==0.7.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.2
onnxruntime==1.16.0
openai==0.28.0
overrides==7.4.0
packaging==23.1
pandas==2.1.0
paramiko==3.3.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.2
pip==23.2.1
posthog==3.0.2
prompt-toolkit==3.0.39
protobuf==4.24.3
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==8.0.0
pyarrow==13.0.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.0
pyEX==0.5.0
Pygments==2.16.1
Pympler==1.0.1
PyMySQL==1.1.0
PyNaCl==1.5.0
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
retry==0.9.2
rich==13.5.2
rpds-py==0.10.2
scipy==1.11.2
setuptools==68.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
socketIO-client-nexus==0.7.6
soupsieve==2.5
SQLAlchemy==2.0.20
sqlalchemy-connector==0.1.42
sqlalchemy_database==0.1.1
sseclient==0.0.27
sshtunnel==0.4.0
stack-data==0.6.2
starlette==0.27.0
streamlit==1.26.0
sympy==1.12
ta-lib-bin==0.4.26
tables==3.8.0
tabulate==0.9.0
temporal-cache==0.1.4
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
typer==0.9.0
typing_extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
validators==0.22.0
watchdog==3.0.0
watchfiles==0.20.0
wcwidth==0.2.6
websocket-client==1.6.3
websockets==11.0.3
wheel==0.38.4
wikipedia==1.4.0
yarl==1.9.2
zipp==3.16.2
